+++ src/main/java/com/codeon/controllers/UserController.java
@@ -1,18 +1,12 @@
 package com.codeon.controllers;
 
-import com.codeon.models.SecurityRole;
-import com.codeon.models.Skill;
-import com.codeon.repositories.SecurityRoleRepo;
-import com.codeon.repositories.SkillRepo;
-import com.codeon.repositories.UserRepo;
-import com.codeon.models.User;
+import com.codeon.models.*;
+import com.codeon.repositories.*;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
 import java.security.Principal;
 import java.util.ArrayList;
@@ -24,12 +18,18 @@ public class UserController {
     private PasswordEncoder passwordEncoder;
     private SecurityRoleRepo securityRoleDao;
     private SkillRepo skillDao;
+    private PostRepo postDao;
+    private PostCommentRepo postCommentDao;
+    private PostTypeRepo postTypeDao;
 
-    public UserController(UserRepo userDao, PasswordEncoder passwordEncoder, SecurityRoleRepo securityRoleDao, SkillRepo skillDao) {
+    public UserController(UserRepo userDao, PasswordEncoder passwordEncoder, SecurityRoleRepo securityRoleDao, SkillRepo skillDao, PostRepo postDao, PostCommentRepo postCommentDao, PostTypeRepo postTypeDao) {
         this.userDao = userDao;
         this.passwordEncoder = passwordEncoder;
         this.securityRoleDao = securityRoleDao;
         this.skillDao = skillDao;
+        this.postDao = postDao;
+        this.postCommentDao = postCommentDao;
+        this.postTypeDao = postTypeDao;
     }
 
     @GetMapping("/register")
@@ -150,5 +150,30 @@ public class UserController {
         user.setSkillList(userSkillList);
         userDao.save(user);
         return "redirect:/login";
+    }
+
+    @GetMapping("/profile")
+    public String profileview(Model model){
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        List<Post> posts = postDao.getAllByUser_Id(user.getId());
+        List<Post> interviewQuestions = postDao.findAllByPostTypeId_Type("interview_question");
+        List<Post> mentorship = postDao.findAllByPostTypeId_Type("mentor_question");
+        List<Post> jobPostings = postDao.findAllByPostTypeId_Type("job_listing");
+        List<PostComment>  postComments = postCommentDao.getAllByUser_Id(user.getId());
+
+        model.addAttribute("jobList", jobPostings);
+        model.addAttribute("mentorList", mentorship);
+        model.addAttribute("postList", interviewQuestions);
+        model.addAttribute("postComment", postComments);
+        model.addAttribute("post", posts);
+        model.addAttribute("user", userDao.getOne((user.getId())));
+        return "users/profile";
+    }
+    @PostMapping("users/{id}/edit")
+    public String profileEdit(@PathVariable long id, @ModelAttribute User user) {
+        User updateInfo = userDao.getOne(id);
+        updateInfo.setBio(user.getBio());
+        userDao.save(updateInfo);
+        return "redirect:/profile";
     }
 }
